Refactor cli clients and cmdexec under parent cli package
